Refactor relevancy unit tests
